L4 Task6  Calculation of the Bill Value
