+++ src/main/java/L4BillCalculator/BillCalculator.java
@@ -6,14 +6,10 @@ public abstract class BillCalculator {
     double billValue;
     float serviceCharge;
 
-
-
-
     public abstract double getBillValueAndServiceCharge();
 
     public abstract double getBillDiscountAndServiceCharge();
 
     public abstract double getTakeAwayCharge();
-
 
 }
